Move client/data/multi-currency/* to multi-currency/client/data/*
