+++ multi-currency/client/data/actions.js
@@ -6,13 +6,14 @@
 import { apiFetch } from '@wordpress/data-controls';
 import { dispatch, select } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
+// TODO: Decouple 'tracks'.
 import { recordEvent } from 'tracks';
 
 /**
  * Internal Dependencies
  */
 import TYPES from './action-types';
-import { NAMESPACE, STORE_NAME } from '../constants';
+import { NAMESPACE, STORE_NAME } from './constants';
 
 export function updateCurrencies( data ) {
 	return {
