+++ multi-currency/client/data/reducer.js
