+++ multi-currency/client/data/selectors.js
